exercise 1.3
The program echo2 is faster than echo1 and echo3 when using 2 strings
I suppose it's because of the for range loop.
strings.Join method uses make and copy for long string, but it's optimized for shorter lenght... might be because it gets the array lenght

The program echo3 is faster than echo1 and echo2 when using more than 3 strings. It's faster at some point to copy byte[] instead of building a string in memory